Added credits. Cleaned up camera view. Adjusted boot splash. Reimported weapon model.
Added credits and linked them to endings and main menu. Cleaned up the camera view to have a shorter clipping distance, as well as preventing the weapon from clipping into the environment. Updated boot splash to final version of "Godotdag". Reimported the weapon model to properly be saved to a scene.